Move files so they no longer need ROS to working (in most cases) Switch to 4-space indentation.
