+++ README.md
@@ -1,4 +1,6 @@
 test_equipment_drivers
 ======================
 
-ROS drivers for various test equipment (DMMs, oscilloscopes, ...) 
+Simple drivers for various test equipment (DMMs, oscilloscopes, ...)
+
+
