--- dso6054a/mainpage.dox
@@ -1,14 +1,0 @@
-/**
-\mainpage
-\htmlinclude manifest.html
-
-\b dso6054a 
-
-<!-- 
-Provide an overview of your package.
--->
-
--->
-
-
-*/
