--- dso6054a/src/dso6054a/__init__.py
@@ -1,8 +1,0 @@
-""" fuctions for capturing data from oscilliscope """
-
-
-    
-    
-    
-
-
