--- futek_usb/src/futek_usb/__init__.py
@@ -1,1 +1,0 @@
-""" Futek USB210 Driver Package """
