+++ inductance.py
