--- loadstar/manifest.xml
@@ -1,16 +1,0 @@
-<package>
-  <description brief="loadstar">
-
-     loadstar
-
-  </description>
-  <author>Derek King</author>
-  <license>BSD</license>
-  <review status="unreviewed" notes=""/>
-  <url>http://ros.org/wiki/loadstar</url>
-  <depend package="rospy"/>
-  <depend package="std_msgs"/>
-
-</package>
-
-
