+++ omega_hh147.py
