--- omega_hh147/Makefile
@@ -1,1 +1,0 @@
-include $(shell rospack find mk)/cmake.mk
