--- scpi_lxi/Makefile
@@ -1,1 +1,0 @@
-include $(shell rospack find mk)/cmake.mk
