--- scpi_lxi/manifest.xml
@@ -1,19 +1,0 @@
-<package>
-  <description brief="scpi_lxi">
-
-    Interactive tool and python library module for communicating with
-    devices that LXI (LAN eXtensions for Instrumentation) interface.
-
-    Newer high-end test equiment (DMM, oscilliscope, power supplies) will 
-    often have a LXI interface.
-
-  </description>
-  <author>Derek King</author>
-  <license>BSD</license>
-  <review status="unreviewed" notes=""/>
-  <url>http://ros.org/wiki/scpi_lxi</url>
-  <depend package="rospy"/>
-
-</package>
-
-
