--- scpi_lxi/src/scpi_lxi/__init__.py
@@ -1,1 +1,0 @@
-""" SCPI LXI driver module """
